Add files via upload
I've been thinking about Tristan Tzara's 'How to Make a Dadaist Poem' (https://www.writing.upenn.edu/~afilreis/88v/tzara.html) and the nature of the process struck me as fairly algorithmic. So I figured trying to code it in Python would be a natural step. I used the standard texts included with the NLTK toolkit in Python, and for Version 1, used 3 of the texts. I plan to extend this for different sorts of texts, and extend the code to include a Sentiment Analysis as well.